Atributo duracao de midias, alterado para inteiro e exclusao de duplicidade de classes 'catalogo'

+++ src/model/Audiobook.java
@@ -3,7 +3,7 @@ package model;
 import enums.Genero;
 
 public class Audiobook extends Midias{
-    public Audiobook(String titulo, String artista, Double duracao, Genero genero) {
+    public Audiobook(String titulo, String artista, Integer duracao, Genero genero) {
         super(titulo, artista, duracao, genero);
     }
