+++ src/model/Midias.java
@@ -5,15 +5,17 @@ import enums.Genero;
 public abstract class Midias {
     protected String titulo;
     protected String artista;
-    protected Double duracao;
+    protected Integer duracao;
     protected Genero genero;
 
-    public Midias(String titulo, String artista, Double duracao, Genero genero) {
+    public Midias(String titulo, String artista, Integer duracao, Genero genero) {
         this.titulo = titulo;
         this.artista = artista;
         this.duracao = duracao;
         this.genero = genero;
     }
+
+
 
     public void exibirInfo() {
         System.out.println(titulo + " - " + artista + " | " + "\n" + duracao + " " + genero);
