+++ src/model/Musica.java
@@ -3,7 +3,7 @@ package model;
 import enums.Genero;
 
 public class Musica extends Midias{
-    public Musica(String titulo, String artista, Double duracao, Genero genero) {
+    public Musica(String titulo, String artista, Integer duracao, Genero genero) {
         super(titulo, artista, duracao, genero);
     }
