+++ src/model/Podcast.java
@@ -3,7 +3,7 @@ package model;
 import enums.Genero;
 
 public class Podcast extends Midias{
-    public Podcast(String titulo, String artista, Double duracao, Genero genero) {
+    public Podcast(String titulo, String artista, Integer duracao, Genero genero) {
         super(titulo, artista, duracao, genero);
     }
